+++ package.json
@@ -3,6 +3,7 @@
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
+  "type": "module",
   "scripts": {
     "test": "jest",
     "build": "serverless package",
@@ -16,6 +17,8 @@
     "osmosis": "^1.1.10"
   },
   "devDependencies": {
+    "@babel/preset-env": "^7.12.11",
+    "babel-jest": "^26.6.3",
     "jest": "^26.6.3",
     "nock": "^13.0.5",
     "serverless": "^1.58.0",
@@ -24,6 +27,9 @@
   "jest": {
     "setupFiles": [
       "<rootDir>/jest/setEnvVars.js"
-    ]
+    ],
+    "transform": {
+      "^.+\\.jsx?$": "babel-jest"
+    }
   }
 }
